backend/requirements.txt
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
gunicorn>=21.2.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
pgvector>=0.2.5

# Redis & Celery
redis>=5.0.1
celery>=5.3.6

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<4.1.0  # Pin to avoid passlib incompatibility with bcrypt>=4.1

# Validation
pydantic>=2.4.1,<2.6
pydantic-settings>=2.1.0
email-validator>=2.1.0

# AI
google-generativeai>=0.8.3

# Telegram Bot
aiogram>=3.4.1,<3.5

# HTTP Client (for GreenAPI)
httpx>=0.27.0

# Voice transcription
elevenlabs>=1.0.0

# Utilities
python-dotenv>=1.0.1
tenacity>=8.2.3


# Search
duckduckgo-search>=5.0.0

# Timezone
pytz>=2024.1

# SQLite async driver
aiosqlite>=0.20.0

# Error monitoring
sentry-sdk>=1.40.0
